requirements_api.txt
fastapi==0.103.1
pydantic==2.4.2
pydantic_core==2.10.1
uvicorn[standard]==0.23.2
python-jose
attrs==23.1.0